chore(authentication): clean up
