+++ lambdas/authentication/utils/token.py
@@ -1,7 +1,31 @@
 import jwt
 from cryptography.fernet import Fernet
 
+import boto3
+
 from config import Config
+
+
+# DynamoDB
+dynamodb = boto3.resource('dynamodb')
+access_token_table = dynamodb.Table(f'{Config.ENVIRONMENT}-AccessToken')
+
+
+def create_access_token_record(company_id, access_token):
+    """
+    Creates a new access token record in the database.
+
+    :param company_id: Unique identifier for the company associated with the token.
+    :param access_token: The access token to be saved.
+    """
+    access_token_info = {
+        'token_id': access_token,
+        'company_id': company_id
+    }
+    try:
+        access_token_table.put_item(Item=access_token_info)
+    except Exception as e:
+        raise RuntimeError(f"Error saving to access token table: {e}")
 
 
 def generate_access_token(company_id):
@@ -22,4 +46,3 @@ def encrypt_github_access_token(access_token: str) -> bytes:
     key = Config.GITHUB_FERNET_KEY
     cipher_suite = Fernet(key.encode())
     return cipher_suite.encrypt(access_token.encode('utf-8'))
-
